feat(sql): use joined strategy as default for SQL drivers (#4958)
This builds on the foundation of #4957, which resolved the biggest
difference of the strategies. As part of the switch, few other smaller
bugs were also fixed, as most of the test suite is now using the joined
strategy too, so it uncovered some bugs that were fixed only with the
select-in.

To keep the old behaviour, you can override the default loading strategy
in your ORM config.