feat(entity-generator): detect more ManyToMany relations
Auto increment columns in pivot entities are now detected.
They are also now properly emitted.

Corrected FK index selection to prefer indexes that match the columns exactly,
and never associate the index with more columns to the property itself
(instead using it at the entity level).
This in turn ensures correctly identifying more ManyToOne relations,
which in turn ensures correctly identifying more ManyToMany opportunities.

In addition, by default, pivot tables are allowed to contain additional props,
including relations, but only if those props would not hinder the ORM's ability
to freely insert and remove records from the collection.

In other words, those additional props need to be optional,
and have defaults that are either null or non-unique.

This can be adjusted with the two new settings
onlyPurePivotTables and readOnlyPivotTables.

Also optimized some of the tests a bit, to never drop the schema,
and create it and its tables only if it doesn't exist already.
The tests trust that if the schema exists, it will be in the expected state.
The DB containers can be re-created if those schemas are changed.