fix(core): mark `Reference.set()` as private
BREAKING CHANGE:

`Reference` wrapper holds an identity—this means its instance is bound to the underlying entity. This imposes a problem when you try to change the wrapped entity, as the same instance of the `Reference` wrapper can be present on other places in the entity graph. For this reason, the `set` method is no longer public, as replacing the reference instance should be always preferred.

```diff
-book.author.set(other);
+book.author = ref(other);
```

Closes #5003