feat(core): allow configuring filters in `Reference.load()` and `Collection.load()`
`Reference.load()` (and other methods that are using `WrappedEntity.init()` under the hood) now return `null` when the target entity is not found instead of resolving to unloaded entity. This can happen either because it was removed in the meantime, or it is not compatible with the currently enabled filters. A new method called `loadOrFail()` is added to the `Reference` class which always returns a value or throws otherwise, just like `em.findOneOrFail`.

This PR also adds more options to both `Reference.load` and `Collection.load`, aligning them with the underlying methods used (`em.findOnd` and `em.populate` respectively).

BREAKING CHANGE

Closes #4975