fix(fs): fixed usages of writeFile(Sync)
Added "flush: true" to ensure that all files written are immediately accessible.
The code already collects the entire contents of a file to write it in one go,
and promises are used to prevent blocking the event loop.

This in turn fixes potentially flaky tests, though the cost is that the overall
time until all file writing promises are resolved will be a bit more.

Also reordered imports in the affected files alphabetically.