refactor: use `flush` option with `writeFile` (#5058)
In all places of the code, fs-extra's writeFile is used. This package
however is actually a re-export of the callback version of node:fs'
writeFile.
Meanwhile, all packages seem to assume this is the version from
node:fs/promises.

This fix in turn fixes tests that can sometimes be flaky, due to
assuming file operations are done, when that isn't necessarily true.

Also reordered imports in the affected files alphabetically.


---

I noticed this while running tests locally. Random tests sometimes fail,
and all they had in common was the fact that they were checking for a
file's existence. I thought it's maybe an fs-extra specific issue with
Windows, but then discovered it's not re-exporting the promise, hence
this PR.

---------

Co-authored-by: Martin Adámek <banan23@gmail.com>