feat(core): auto-join M:1 and 1:1 relations with filters (#5063)
Filters are applied to the relations too, as part of `JOIN ON`
condition. If a filter exists on a M:1 or 1:1 relation target, such an
entity will be automatically joined, and when the foreign key is defined
as `NOT NULL`, it will result in an `INNER JOIN` rather than `LEFT
JOIN`. This is especially important for implementing soft deletes via
filters, as the foreign key might point to a soft-deleted entity. When
this happens, the automatic `INNER JOIN` will result in such a record
not being returned at all.

You can disable this behavior via `autoJoinRefsForFilters` ORM option.

Related: #4975