fix(core): allow calling `em.remove` with not managed entity
This breaking change was introduced via 1b09d263f94f4c216936f6f85673810f7c6f9099, but we wanted to ignore the `em.remove` only in case previous `em.persist` call was there, the current implementation could cause hard to debug issues as it silently ignores the `remove` call in that case.

This commit makes it possible to call `em.remove` on not managed entities again, as long as they are not part of the persist stack.

Closes #5103