test(timers): replaced all timeouts with fake timers, or remove them as appropriate (#5140)
Also tweaked the test entities-js/Author3.js and
entities-schema/BaseEntity5.ts to derive the update out of the create or
(in the case of the entity schema) vice-versa. This refactor, combined
with fake timers in the affected tests, means that no tolerance is
required for those columns - they start off as equal, always.