feat(core): allow mapping array columns to arrays of objects via `ArrayType` (#5204)
`ArrayType` can be now also used to map the array items to more complex
types like objects. Consider the following example of mapping a `date[]`
column to array of objects with `date` string property:

```ts
import { ArrayType } from '@mikro-orm/core';

export class CalendarDateArrayType extends ArrayType<CalendarDate> {

  constructor() {
    super(
      date => ({ date }), // to JS
      d => d.date, // to DB
    );
  }

  getColumnType(): string {
    return 'date[]';
  }

}
```

This PR also disables the automatic conversion of arrays of datetime
values (and similar, e.g. date or interval) to arrays of Date objects.
This was disabled only for the scalar types, but it had to be disabled
explicitly for their array variants to get consistent behavior.

Closes #5188