docs: updated dependencies (#5248)
Also, now that no broken links are reported, modified the config to
throw if any are introduced.

Note that the version of @mdx-js/react is fixed to 3.0.0, as 3.0.1 seems
to break the anchor collection in docusaurus' mdx-loader. Not sure if
docusaurus is interacting with it erroniously or if @mdx-js/react
unintentionally broke what used to work, or if maybe one of
@mdx-js/react's dependencies broke something, so I haven't reported this
to either repo yet.


----

Not sure why renovate still didn't pick up the new plugin version. My
best guess is because it's not part of the root's workspace. If
successfully enabled, it might try to update @mdx-js/react, and right
now, it will probably fail due to the new version introducing broken
anchors.