fix(entity-generator): optional and hidden properties get type option + string defaults (#5264)
The "default" option is no longer emitted into the decorator, in favor
of it being specified in the property's declaration at the entity class.
The "defaultRaw" option is still used in cases where the SQL expression
does not unambiguously map to the "default" option's value.

The MySQL schema helper translates TINYINT(1) with a value of 0 to
"false". This ensures behavior consistent with other drivers. And if
this value accidentally finds its way back in a query, it is also OK, as
MySQL treats this literal as "0".

The "optional" option is now added early, in the DatabaseTable stage.
This has also unlocked some ManyToMany relations that were previously
not detected.

Closes #5260