perf(core): optimize adding entities to large collections
Skips checking existing collection items when looking for EM, which was happening on each `add` call due to it canceling the orphan removal automatically. This check shouldn't be needed, as all the items that are added to the collection will be checked already, so we would have a cached EM reference once it's found somewhere in the input entities.