fix(sql): override FK value when it's disallowed by query condition
Previously, a FK value was always hydrated based on the owning entity, and if its target entity was not compatible with current query condition, such relation couldn't be populated. With this change, the PK value of the joined relation is preferred instead of the FK value, meaning that if we fail to join the entity, we hydrate `null` in place of the actual FK value. This is in line with how filters behave since v6.