docs(core): added definition of an identity map (#5369)
Related to #5367 .

The sentence
> they don't really explain anything to someone not knowing what an
Identity Map is.

kind of triggered me on this one, in that it is actually a fair point,
despite not being the core of the issue. Not sure if my definition here
makes sense, but I think _some_ definition of the term identity map
being present is important.