feat(schema): improve `orm.schema.execute()` to support executing batches
When you run a batch of queries via `schema.execute()`, they are now grouped based on blank lines - a blank line between groups of statements means they will be executed separately. This is important for some queries that cannot be ran in a single batch, like when you create a database and switch to it.