feat(core): make `assign` options configurable globally
Also changes how assigning to object properties work once again. This was changed in v5 with a refactoring related to embedded entities, but mistakenly propagated to the JSON properties too.

Now we do not merge data when assigning to JSON properties by default, and require explicit `mergeObjectProperties: true` in the options. For assigning to embedded properties, this does not apply, and a new option called `mergeEmbeddedProperties` is introduced to allow disabling the default behavior.

Closes #5410