fix(core): detect path from decorator for each class only once (#5545)
### Fix for bun path detection
Related discussion here:
https://github.com/mikro-orm/mikro-orm/discussions/5496
In [this
commit](https://github.com/mikro-orm/mikro-orm/commit/271ff7dfe95ded8f2821b7a97a10ddfe409b1513)
you changed the condition because of coverage. This is incorrect because
it guards against going out of stack array bounds - I have added the
test for this case. Also, I removed no longer needed condition.

### Performance
Before the changes, the path was detected for every used decorator on
the class and overwritten. This is not needed because the path used as a
key should be stable and should not change for a given class. So I
introduced the new symbol `MetadataStorage.PATH_SYMBOL`, which will hold
the path and check if it exists for a given class using `Object.hasOwn`
(this also works for class inheritance so the path is not taken from the
base class but defined for each child).

Moreover, as a side effect, this also allows defining a path on the
class by the developer (maybe in cases like when Bun was not working),
but it should not be needed and used only as a workaround:

```typescript
import { Entity, MetadataStorage } from '@mikro-orm/core';

@Entity()
class User {
  static [MetadataStorage.PATH_SYMBOL] = import.meta.filename // or __filename for commonjs

  @PrimaryKey()
  id!: number;

  @Property()
  username!: string;
}
```