perf(sql): optimize diffing M:N collection state
In this particular case, the diffing added ~800ms that were caused by a comparison of 2000 collections represented by their PKs in the form of arrays.
The comparison was done via `Utils.equals()` which is completely generic. Now we use an optimized version of `compareArray()` method, which checks scalar
values by `===` instead. As a result, the diffing went from ~800ms down to ~30ms.

Related: #5627