fix(knex): explicitly declare all the extended drivers as optional peer dependencies
This help avoid bundlers like Vite trying to pre-package them,
and instead leaves them as externals, without the need of an explicit config change.

Note that Knex itself declares all of the drivers it supports as optional peer dependencies.
This package just defines its extensions as optional peer dependencies in addition to that.