feat(entity-generator): allow custom types for scalar relations (#5435)
This is done by fully refactoring type inference in general,
to take advantage of "runtimeType" in the prop,
combined with more cross checks to add IType where needed.

Also removed the accidental Embedded decorator in EntitySchema.

---

Using custom types for scalar relations was sort of possible even before
this, but it has always been more hacky than anything else. With this,
it is actually encoded as an expected possibility, and added to the test
for hooks. Users who want to use custom types will need to handle their
resolution in the fileName function. Same as if those identifiers were
referring to an entity.

Co-authored-by: Martin Adámek <banan23@gmail.com>