fix(entity-generator): fix handling of primary keys that are foreign keys or enums (#5673)
Previously, PKs that are not scalar would miss the "primary" option in
the output, and enums would use the PrimaryKey decorator instead of the
Enum decorator.

Also, FKs that are refs to enum columns would previously create a brand
new enum class in each ref. Now, only the "origin" column gets the enum
class, and other references use the entity class.