fix(sqlite): fix altering tables via malformed temp table (#5683)
This replaces the knex implementation which was dependent on the current
state, which was wrong on its own, since the state might not be matching
the metadata we have (e.g. when generating migration against a
snapshot). Now we use the exact same `create table` statement as when
we create the schema from scratch.

Closes #5672