fix(mssql): add proper support for MSSQL's native "varchar" type
Previously, specifying "type: 'varchar'" on a property was treated as aliasing it to "nvarchar"
for schema diffing and entity generation purposes,
without wrapping it up as UnicodeString at runtime.

Now, it is using MSSQL's native "varchar" for schema diffing / entity generation purposes.
Those values are not wrapped up in the UnicodeString wrapper at runtime, as before.

For other drivers, "varchar" still means "varchar if supported" (sqlite still aliases to "text").