fix(core): respect nullability of wrapped types for ScalarReference's (#5722)
If one is to use a ScalarReference to get type-safe access to a lazily
loaded scalar value it should be possible to encode in the types that
the scalar may be null in the database. Currently, ScalarReferences,
when de-referenced through `.get()` or `.$` are typed with "NonNullable"
variants of their wrapped types. That non-nullability is not reflected
at runtime, which opens up code that perfectly type-checks to runtime
exceptions because of null reference accesses. This is demonstrated in
this reproduction: https://github.com/ropbastos/mikro-orm-reproduction

Briefly linking the "fixed" version of the library with my companies
project I didn't see any types that looked wrong, and all tests still
pass, so I'm creating this PR. However, I'm aware the NonNullable
helpers are pretty explicit about their intent 😅. I'm not sure what I'm
trying to do makes sense/why were they added in the first place, but in
the off-chance that they are there by mistake (or they ceased to make
sense), I've already done the small work of removing them 🙏

Here is a discussion I opened when I first encountered an error of the
type represented in the reproduction repository:
https://github.com/mikro-orm/mikro-orm/discussions/5711

---------

Co-authored-by: Martin Adámek <banan23@gmail.com>