feat(core): added Platform.getDefaultVarcharLength and optional Type.getDefaultLength
The existing types are converted to rely on this method's presence.

The entity generator now omits the length property when it is equal to the default.

The schema generator will not detect a difference between not having length
(when that method is defined on the type) and having the length with the default value.
As always, specifying columnType bypasses all of that.