feat(migrations): allow initial migration to be blank if no entities are defined
This enables easier "schema first" approach, whereby users first create the schema
through an initial migration, before generating the entities out of it.

The "No entities found" error is still present when the "--blank" flag is absent.

When using "--blank" and "--initial" together, the new blank migration is not logged.