fix(entity-generator): unknown type defaults always use default/defaultRaw, never runtime (#5927)
The type may be ignored for schema changes if it features length, since
we can not guarantee that the "length" is the first argument (if at all
present), yet is what the generated columnType assumes.
If the type is ignored for that reason, the "extra" (if present) will
also be ignored. Metadata hooks may remove the ignoreSchemaChanges
option as a way to confirm it is OK.

The default values of unknown types are left unwrapped in MySQL and
MSSQL, which leaves them written as defaultRaw.