fix(core): fix same-second dates diffing (#6094)
This small PR fixes a minor regression that I think was introduced in
https://github.com/mikro-orm/mikro-orm/commit/55df57ff1aa84e5d45188d849ba09e91ae6d3642.

## Issue
When updating an entity and changing a date property to another one very
close to it (by a difference in microseconds), the change is not
persisted in the database. Indeed, diffing dates by comparing the
`toString()` output doesn't give enough precision to compare
microseconds shifts. Example:
- `toString()` output: `Tue Oct 01 2024 08:54:48 GMT+0000 (Coordinated
Universal Time)`
- `toISOString()` output: `2024-10-01T08:54:48.651Z`

Two dates can have the same `toString()` output even though their
`toISOString()` output is different.

The impact is low and probably only concerns the test scenarios: It's
rare to persist the same entity twice in the same second without
changing anything other than a date.

## Fix
In this PR I'm switching the dates comparison to use `toISOString()` and
adding a new unit test.

---------

Co-authored-by: Martin Adámek <banan23@gmail.com>