feat(entity-generator): added support for forceUndefined being false (#5849)
There's a separate option "undefinedDefaults" about whether the reported
null default values should be changed to "undefined", and those columns
be made optional, as opposed to require null.

Also fixed empty string defaults in either mode.

The Opt type modifier is no longer added to properties with defaults in
the property declaration. It is still added when the "default" is only
at the option, and not at the property declaration. Also not added to
properties with "?" type annotation.