refactor(cli): made the CLI explicitly pass the config path from the arg, or else infer it once
The output of debug is adjusted as a consequence of the config being a passed dependency.
Driver dependency versions are going to be shown if a config is found,
independently of whether a package.json was found.

The config is always validated, even during entity generation.
When warnings about missing entities are disabled,
the entities options are also set to empty arrays,
to avoid the old / additional entities' meta data influencing the entity generation.

If the config path is automatically inferred from a CLI argument during MikroORM.init(),
a deprecation warning is logged.

checkConnection are adjusted to be type safe,
i.e. "reason" only exists on error,
and "error" may potentially exist on error, but never when "ok" is true.