feat(core): upsert managed entities
Previously, if you tried to upsert a managed entity, the call was interpreted as `assign`, and an explicit `flush` was required to persist those changes to the database. Now the upsert query is always issued regardless of the entity being managed or not.

To opt in to the previous behavior, use `upsertManaged: false` in the ORM config.

Closes #6055