feat(core): added the ability to define multiple named configurations
A config file may now export a function or an array.

When a function is exported, it is called with a runtime supplied configuration name
("default" by default), and is expected to return a config object,
or a promise resolving to a configuration object.

When an array is supplied, it is searched for a config object
with a "name" option matching the user supplied one.
Not having that option set is the same as setting it to "default".

If the array does not contain a config with that name, or the function returns undefined,
the config file is skipped, and the next possible ones are searched.
If config files were found, but none of them gave the desired name config,
a new error detailing the found config files is thrown.