fix(core): overloaded ConfigurationLoader.getConfiguration() to make it cross version compatible
If true is explicitly given at the first argument, it will trigger the legacy behavior,
where `--config` is also searched for in `process.argv`.

If left without arguments, the new default is used, which is still mostly compatible,
except `--config` is not searched in `process.argv`.

In either situation, the paths to search for default to
the result of `ConfigurationLoader.getConfigPaths()`.