refactor: the BC layer of ConfigurationLoader.getConfiguration now fully emulates old behavior
There is now a search for files even on validate files, so that if there is a config,
it can be augmented with envs rather than an env only config being returned.

The search for files itself is now moved into a separate helper function called getConfigFile
which returns a tuple of path and config file contents (whatever it may be)
or an empty tuple (when none of the config files existed or failed to be imported).
As a nice side effect of this,
the complexity of ConfigurationLoader.getConfiguration is reduced by a little.