fix(core): prevent reinsertion of an entity referenced by a formula when it is removed from a collection (#6265)
### The Problem

When using a `toMany` relation on an entity and also adding a `formula`
that references the same entity in the collection, removing the entity
from the collection causes the `formula` relation to keep referencing
it, resulting in the deleted entity being reinserted on flush.

### Description of the Solution

I updated `referencingProperties` to allow multiple references on the
same model, not just the first. However, the problem persisted because
even though `unsetIdentity` performs `delete
helper(referrer).__data[prop.name];`, it does nothing since `__data` is
an empty object. I added an explicit override of the reference with
`referrer[prop.name] = null;`.

I’m not sure what the correct way to reset this reference is (another
test is failing because of that), so advice would be appreciated.
Options include:
- Setting it to `null`  
- Setting it to `undefined`  
- Deleting the property with `delete referrer[prop.name]`

---------

Co-authored-by: Martin Adámek <banan23@gmail.com>