fix(query-builder): disallow awaiting the QueryBuilder instance (#6438)
BREAKING CHANGES:
- QueryBuilder is no longer awaitable, it can be now returned from an
async method, which previously triggered the promise. Use `qb.execute()`
or `qb.getResult()` instead of awaiting it directly.
- `@CreateRequestContext` requires async function. The decorator
converts the function to be async, now it will trigger a validation
error if the function it is applied to is not async, since it could
result in an ignored promise.

Closes #4876