feat(entity-generator): pass through `orderBy` option to the output (#6466)
I am using entity generator to maintain source code of my entities, and
also I'd like to provide `orderBy` option for some of my collections.

Currently one can alter a lot of entities' code through the
`onProcessedMetadata` callback - but just not the `orderBy` option, as
it is never saved to the generated code output.

This PR fixes it - allowing to specify `orderBy` option for some of the
relations in the metadata callbacks, so that it's later placed in the
generated entities code.