fix(core): respect Collection property `orderBy` when dataloader is enabled (#6467)
After adding default `orderBy` to my collection in the entity class
`@OneToMany(...)` decorator, I noticed that my collections are sometimes
sorted correctly and sometimes not at all.

Upon debugging it appeared that this happens only if `dataloader` is
enabled (and it is, in my case).

This PR fixes it, making the `Collection#loadItems` load items in the
default order (rather than none) also when dataloader is enabled.
(Currently, it happens only when it is disabled.)