fix(core): skip TS support detection via `require.extensions` (#6488)
This check was problematic because of Node.js type stripping uses it now
too, and it does fail to parse decorators.

It was already removed from v7, but since the type stripping will be
backported to Node.js 22, I am removing it from v6 too, otherwise things
break when executing the compiled code.

Related to
https://github.com/nodejs/node/pull/57298#issuecomment-2703430792