feat(core): add `includeCount` option to `em.findByCursor` (#6536)
This PR restores the changes from the original PR #4626 (which were
possibly lost due to a bad rebase) and resolves merge conflicts with the
current master branch.

As B4nan suggested in the original PR, I've made the changes to the
Cursor constructor parameters non-breaking, so the feature can be
released under the current major version.

P.S. Thanks for the awesome ORM! :)

---------

Co-authored-by: Martin Adámek <banan23@gmail.com>