refactor(core): change default of `embeddables.prefixMode` to `relative`
BREAKING CHANGE:

Earlier versions always used the `absolute` prefix mode for nested embedded properties with overridden `columName`, meaning the column name set on the scalar property level ignored the parent `prefix` option. This was technically a bug, which was resolved in v6.4.0, but to stay backwards compatible, the default behavior stayed the same (`prefixMode: 'absolute'`). Since v7, the `prefixMode` defaults to `relative`, which means the `columnName` will be prefixed based on the embedded property `prefix` option (and its parents).