fix(core): map arrays inside object embeddables to JSON arrays by default
BREAKING CHANGE:

Previously, array properties inside object embeddables were not automatically considered as JSON properties—instead, they mapped to the `ArrayType` as a regular top level property would have. This means that the value in the database was stored as a comma separated list (or an array literal in postgres). In v7, such properties are considered as JSON arrays automatically.

If you want to preserve the previous behavior, set the type explicitly:

```diff
-@Property()
+@Property({ type: ArrayType })
 array!: string[];
```