fix(schema): properly handle array defaults in mysql (#6703)
Fixes #6688. While testing across all databases, MySQL was failing with
the following error:

```sql
DriverException: create table `a` (
  `id` int unsigned not null auto_increment primary key,
  `languages` text not null default 'en',
  `accounts` json not null default '[]',
  `items` text not null default ''
) default character set utf8mb4 engine = InnoDB;
- BLOB, TEXT, GEOMETRY or JSON column 'languages' can't have a default value
```
See: https://dev.mysql.com/doc/refman/9.0/en/data-type-defaults.html

> The [BLOB](https://dev.mysql.com/doc/refman/9.0/en/blob.html),
[TEXT](https://dev.mysql.com/doc/refman/9.0/en/blob.html), GEOMETRY, and
[JSON](https://dev.mysql.com/doc/refman/9.0/en/json.html) data types can
be assigned a default value only if the value is written as an
expression, even if the expression value is a literal

Fixed by wrapping the default value in an expression when the column
type is one of the types mentioned above.