feat(schema): allow wilcard-schema friendly index expression (#6706)
This feature allows to use index expressions on wilcard-schema entities.

Now `expression` accepts either a SQL expression as a raw `string` (aka
same behavior as til now); or a callback function that returns the SQL
expression to be used, the dynamic schema being passed as a parameter.

```ts
@Entity({ schema: '*' })
@Index({ expression: (schema?: string) => `create index "custom_idx_on_name" on "${schema}"."author" ("name")` })
export class Author {
  ...
  @Property()
  name: string;
  ...
}
```

Logic is as follow:
- If the entity is setup with schema: "my_schema", "my_schema" is passed
as a parameter.
- If the entity is setup with schema: "*", the dynamic schema is passed
as a parameter (aka the targeted schema while calling
`schemaGenerator.getCreateSchemaSQL({ schema: "..." })`).

Closes #6446