chore: added healthcheck commands for all DB engines (#6888)
Use of TCP is explicitly encoded into all commands, ensuring they don't
connect successfully until TCP is available, even if shared memory
already is.

For MSSQL, the "-b" flag is also added,
which ensures failures are registered as a status code change, not just
an error message, which is what happens without it.