feat(core): allow defining filters on nullable relations as strict (#6952)
The behavior introduced by the fix for #6784 turned out as breaking a
common use case for filters, which is soft-deletes. It turns out,
filters on relations need to behave differently depending on the use
case. Sometimes you want to discard the owning entity because its
nullable relation is discarded by filter (e.g. for a tenant filter),
while for soft deleted relation, you'd want to hydrate the discarded
relation to null instead of discarding the owning entity.

This PR adds a new `strict` option to the filter definition, which
controls exactly this.

Closes #6943
Closes #6784