feat(core): support private properties with get/set accessors (#6967)
When using a private property backed by a public get/set pair, use the
`accessor` option to point to the other side.

> The `fieldName` will be inferred based on the accessor name unless
specified explicitly.

If the `accessor` option points to something, the ORM will use the
backing property directly:

```ts
@Entity()
class User {
  // the ORM will use the backing field directly
  @Property({ accessor: 'email' })
  private _email: string;

  get email() {
    return this._email;
  }

  set email(email: string) {
    return this._email;
  }
}
```

If you want to the ORM to use your accessor internally too, use
`accessor: true` on the get/set property instead. This is handy if you
want to use a **native private property** for the backing field.

```ts
@Entity({ forceConstructor: true })
class User {
  #email: string;

  // the ORM will use the get/set field internally
  @Property({ accessor: true })
  get email() {
    return this.#email;
  }

  set email(email: string) {
    return this.#email;
  }
}
```

Closes #6932