feat(cli): support various TS loaders (#6987)
Adds native support for various TS loaders for the TS support in the
CLI:
- `swc` via `@swc-node/register`, supports metadata reflection
- `tsx`
- `jiti`
- `tsimp`

The default is `auto`, which means it goes through all those options
sequentially and picks the first one available in the project
dependencies.

To pick a loader explicitly, use the `tsLoader` setting in your
package.json:

```json
"mikro-orm": {
  "tsLoader": "jiti"
}
```

Or override it via `MIKRO_ORM_CLI_TS_LOADER` env var.