perf(core): simplify processing of scalars that use simple ORM provided type classes
When a custom type is used in entity definition, it results in more complex hydration and diffing code, since we need to invoke the type methods dynamically. This is adding unnecessary perf penalty when simple types with `ensureComparable: false` are processed. This PR rewrites the type option in the metadata to a simple type name in favor of the type class, to have simpler hydration code generated.