refactor: remove validate and strict options (#7029)
BREAKING CHANGE:

`validate` and `strict` options are removed.

Both are now enabled by default, and the auto-fixing mechanism is
removed. This means that if you try to map a raw result from the
database, it needs to be correctly typed. One example where this can
happen is when you use some aggregation function like `sum`, in some
dialects like postgres, it produces strings by default, which are no
longer mappable to a `number` property by default.