fix(defineEntity): avoid circular type inference in meta (#7039)
This PR fixes the type circular error that occurs when using
`defineEntity` to define a circular entity, for properties such as
`indexes`, `uniques`:

```ts
const Category = defineEntity({
  name: 'Category',
  uniques: [
    { properties: ['name', 'children'] },
  ],
  properties: {
    id: p.integer().primary(),
    name: p.string(),
    description: p.string().length(2048).nullable(),
    metadata: p.json<Record<string, any>>().default('{}'),
    isShopEnabled: p.boolean(),
    parentCategory: () => p.manyToOne(Category).nullable(),
    children: () => p.oneToMany(Category).mappedBy('parentCategory'),
  },
});
```

The cause of this problem is that `EntityKey<T>` reads the property
values of the circular entity type, and TypeScript starts to compute the
specific values of the circular entity type.
The solution is to use `keyof TProperties` instead of `EntityKey<T>` in
`defineEntity`.